create example project in example folder
