--- example/main.dart
@@ -1,91 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:motion_tab_bar/MotionTabBarView.dart';
-import 'package:motion_tab_bar/MotionTabController.dart';
-import 'package:motion_tab_bar/motiontabbar.dart';
-
-void main() => runApp(MyApp());
-
-class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Motion Tab Bar Sample',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: MyHomePage(title: 'Motion Tab Bar Sample'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
-  MotionTabController _tabController;
-
-  @override
-  void initState() {
-    super.initState();
-    _tabController = MotionTabController(initialIndex: 1, vsync: this);
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _tabController.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: Text(widget.title),
-        ),
-        bottomNavigationBar: MotionTabBar(
-          labels: [
-            "Account","Home","Dashboard"
-          ],
-          initialSelectedTab: "Home",
-          tabIconColor: Colors.green,
-          tabSelectedColor: Colors.red,
-          onTabItemSelected: (int value){
-            print(value);
-            setState(() {
-              _tabController.index = value;
-            });
-          },
-          icons: [
-            Icons.account_box,Icons.home,Icons.menu
-          ],
-          textStyle: TextStyle(color: Colors.red),
-        ),
-        body: MotionTabBarView(
-          controller: _tabController,
-          children: <Widget>[
-            Container(
-              child: Center(
-                child: Text("Account"),
-              ),
-            ),
-            Container(
-              child: Center(
-                child: Text("Home"),
-              ),
-            ),
-            Container(
-              child: Center(
-                child: Text("Dashboard"),
-              ),
-            ),
-          ],
-        ));
-  }
-}
